requirements.txt
openai-whisper
sounddevice
numpy
psutil
scipy
anthropic
pystray
pillow
truststore
pycaw
